refactor(agent-service): split CRUD god object into per-entity modules
crud.py (316 lines) split into crud/persona.py, crud/model_provider.py,
crud/message.py, crud/schedule.py, crud/life_engine.py.
Consolidated raw SQL from services/workers into CRUD layer:
- afterthought.py, glimpse.py: inline group_name query → crud/message.get_group_name
- glimpse_worker.py: inline life_engine_state query → crud/life_engine.load_latest_state
- vectorize_worker.py: message queries → crud/message
- chat_consumer.py, post_consumer.py: response updates → crud/message
- bot_context.py, message_router.py: persona queries → crud/persona
Re-export __init__.py maintains backward compatibility.
Updated 4 test files to match new module paths.
