+++ apps/agent-service/app/orm/crud/schedule.py
@@ -1,154 +1,9 @@
+"""AkaoSchedule CRUD operations"""
+
 from sqlalchemy.future import select
 
-from .base import AsyncSessionLocal
-from .models import (
-    AkaoSchedule,
-    ConversationMessage,
-    LarkBaseChatInfo,
-    LarkGroupChatInfo,
-    LarkUser,
-    ModelMapping,
-    ModelProvider,
-)
-
-
-def parse_model_id(model_id: str) -> tuple[str, str]:
-    """
-    解析model_id格式："{供应商名称}:模型原名"
-
-    Args:
-        model_id: 格式为"供应商名称:模型原名"的字符串
-
-    Returns:
-        tuple: (供应商名称, 模型原名)
-    """
-    if ":" in model_id:
-        provider_name, model_name = model_id.split(":", 1)
-        return provider_name.strip(), model_name.strip()
-    else:
-        # 如果没有/，使用默认供应商302.ai
-        return "302.ai", model_id.strip()
-
-
-async def get_model_and_provider_info(model_id: str):
-    """
-    根据model_id获取供应商配置和模型名称
-    优先查找 ModelMapping，如果未找到则回退到解析逻辑
-
-    Args:
-        model_id: 映射别名 或 格式为"供应商名称/模型原名"的字符串
-
-    Returns:
-        dict: 包含模型和供应商信息的字典
-    """
-    async with AsyncSessionLocal() as session:
-        # 1. 尝试查找映射
-        mapping_result = await session.execute(
-            select(ModelMapping).where(ModelMapping.alias == model_id)
-        )
-        mapping = mapping_result.scalar_one_or_none()
-
-        if mapping:
-            provider_name = mapping.provider_name
-            actual_model_name = mapping.real_model_name
-        else:
-            # 2. 回退到解析逻辑
-            provider_name, actual_model_name = parse_model_id(model_id)
-
-        # 3. 查询供应商信息
-        provider_result = await session.execute(
-            select(ModelProvider).where(ModelProvider.name == provider_name)
-        )
-        provider = provider_result.scalar_one_or_none()
-
-        # 如果找不到指定供应商，尝试使用默认的302.ai
-        if not provider:
-            provider_result = await session.execute(
-                select(ModelProvider).where(ModelProvider.name == "302.ai")
-            )
-            provider = provider_result.scalar_one_or_none()
-
-        if not provider:
-            return None
-
-        return {
-            "model_name": actual_model_name,
-            "api_key": provider.api_key,
-            "base_url": provider.base_url,
-            "is_active": provider.is_active,
-            "client_type": provider.client_type or "openai",
-            "use_proxy": provider.use_proxy,
-        }
-
-
-async def get_bot_persona(persona_id: str) -> "BotPersona | None":
-    """获取 bot 人设配置"""
-    from app.orm.models import BotPersona
-    async with AsyncSessionLocal() as session:
-        return await session.get(BotPersona, persona_id)
-
-
-async def get_gray_config(message_id: str) -> dict | None:
-    """
-    根据 message_id 关联查询所属 chat 的灰度配置
-    """
-    async with AsyncSessionLocal() as session:
-        # 使用 Join 避免 N+1 查询
-        stmt = (
-            select(LarkBaseChatInfo.gray_config)
-            .join(
-                ConversationMessage,
-                ConversationMessage.chat_id == LarkBaseChatInfo.chat_id,
-            )
-            .where(ConversationMessage.message_id == message_id)
-        )
-        # 直接返回配置字段 (dict) 或者 None
-        return await session.scalar(stmt)
-
-
-async def get_message_content(message_id: str) -> str | None:
-    """
-    根据 message_id 获取消息内容
-
-    Args:
-        message_id: 消息ID
-
-    Returns:
-        消息内容，如果未找到返回 None
-    """
-    async with AsyncSessionLocal() as session:
-        stmt = select(ConversationMessage.content).where(
-            ConversationMessage.message_id == message_id
-        )
-        return await session.scalar(stmt)
-
-
-async def get_chat_messages_in_range(
-    chat_id: str, start_time: int, end_time: int, limit: int = 2000
-) -> list[ConversationMessage]:
-    """获取指定群在时间范围内的所有消息（user + assistant）"""
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(ConversationMessage)
-            .where(ConversationMessage.chat_id == chat_id)
-            .where(ConversationMessage.create_time >= start_time)
-            .where(ConversationMessage.create_time < end_time)
-            .order_by(ConversationMessage.create_time.asc())
-            .limit(limit)
-        )
-        return list(result.scalars().all())
-
-
-async def get_username(user_id: str) -> str | None:
-    """从 lark_user 表获取用户名"""
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(LarkUser.name).where(LarkUser.union_id == user_id)
-        )
-        return result.scalar_one_or_none()
-
-
-# ==================== AkaoSchedule CRUD ====================
+from app.orm.base import AsyncSessionLocal
+from app.orm.models import AkaoSchedule
 
 
 _WEEKDAY_MAP = {
@@ -306,11 +161,3 @@ async def delete_schedule(schedule_id: int) -> bool:
         await session.delete(entry)
         await session.commit()
         return True
-
-
-async def get_all_persona_ids() -> list[str]:
-    """获取所有 persona 的 persona_id 列表"""
-    from app.orm.models import BotPersona
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(select(BotPersona.persona_id))
-        return [row[0] for row in result.all()]
