+++ apps/agent-service/app/services/message_router.py
@@ -6,13 +6,14 @@ Phase 3 扩展点: 无 @ 通用判断器、主动发言路由。
 
 import logging
 
+from app.orm.crud.persona import resolve_mentioned_personas, resolve_persona_id
+
 logger = logging.getLogger(__name__)
 
 
 async def _resolve_persona_id(bot_name: str) -> str:
-    """从 bot_config 表查 persona_id（复用 bot_context 的逻辑）"""
-    from app.services.bot_context import _resolve_persona_id as _resolve
-    return await _resolve(bot_name)
+    """从 bot_config 表查 persona_id（复用 crud 层）"""
+    return await resolve_persona_id(bot_name)
 
 
 class MessageRouter:
@@ -42,7 +43,7 @@ class MessageRouter:
             return [pid]
 
         if mentions:
-            persona_ids = await self._resolve_mentioned_personas(mentions)
+            persona_ids = await resolve_mentioned_personas(mentions)
             logger.info(
                 "Group @mention route: mentions=%s → persona_ids=%s",
                 mentions, persona_ids,
@@ -51,25 +52,3 @@ class MessageRouter:
 
         # 群聊无 @ → 不回复（Phase 3 扩展点）
         return []
-
-    async def _resolve_mentioned_personas(
-        self, mentions: list[str]
-    ) -> list[str]:
-        """将 mention 的 app_id 列表映射到 persona_id 列表
-
-        查询 bot_config 表：mention 的 app_id 匹配 app_id 列。
-        """
-        from app.orm.base import AsyncSessionLocal
-        from sqlalchemy import text
-
-        async with AsyncSessionLocal() as session:
-            result = await session.execute(
-                text(
-                    "SELECT DISTINCT persona_id FROM bot_config "
-                    "WHERE app_id = ANY(:mentions) "
-                    "AND is_active = true "
-                    "AND persona_id IS NOT NULL"
-                ),
-                {"mentions": mentions},
-            )
-            return [row[0] for row in result.fetchall()]
