+++ apps/agent-service/app/workers/glimpse_worker.py
@@ -4,25 +4,10 @@ import logging
 
 from app.config.config import settings
 from app.orm.crud import get_all_persona_ids
+from app.orm.crud.life_engine import load_latest_state
 from app.services.glimpse import run_glimpse
 
 logger = logging.getLogger(__name__)
-
-
-async def _load_life_engine_state(persona_id: str):
-    """查 Life Engine 最新状态"""
-    from app.orm.base import AsyncSessionLocal
-    from app.orm.memory_models import LifeEngineState
-    from sqlalchemy.future import select
-
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(LifeEngineState)
-            .where(LifeEngineState.persona_id == persona_id)
-            .order_by(LifeEngineState.created_at.desc())
-            .limit(1)
-        )
-        return result.scalar_one_or_none()
 
 
 async def cron_glimpse(ctx) -> None:
@@ -36,7 +21,7 @@ async def cron_glimpse(ctx) -> None:
 
     for persona_id in await get_all_persona_ids():
         try:
-            state = await _load_life_engine_state(persona_id)
+            state = await load_latest_state(persona_id)
             if not state or state.activity_type != "browsing":
                 continue
             await run_glimpse(persona_id)
