+++ apps/agent-service/app/workers/vectorize_worker.py
@@ -13,8 +13,6 @@ import uuid
 
 from aio_pika.abc import AbstractIncomingMessage
 from inner_shared.logger import setup_logging
-from sqlalchemy import update
-from sqlalchemy.future import select
 
 from app.agents import InstructionBuilder, create_client
 from app.clients.image_client import image_client
@@ -25,7 +23,11 @@ from app.clients.rabbitmq import (
     _lane_queue,
 )
 from app.clients.redis import AsyncRedisClient
-from app.orm.base import AsyncSessionLocal
+from app.orm.crud.message import (
+    get_message_by_id,
+    scan_pending_messages as _crud_scan_pending,
+    update_vector_status,
+)
 from app.orm.models import ConversationMessage
 from app.services.download_permission import check_group_allows_download
 from app.services.qdrant import qdrant_service
@@ -56,28 +58,6 @@ def _handle_signal(signum, frame):
     global _running
     logger.info(f"收到信号 {signum}，准备优雅退出...")
     _running = False
-
-
-async def get_message_by_id(message_id: str) -> ConversationMessage | None:
-    """从数据库获取消息"""
-    async with AsyncSessionLocal() as session:
-        result = await session.execute(
-            select(ConversationMessage).where(
-                ConversationMessage.message_id == message_id
-            )
-        )
-        return result.scalar_one_or_none()
-
-
-async def update_vector_status(message_id: str, status: str) -> None:
-    """更新消息的向量化状态"""
-    async with AsyncSessionLocal() as session:
-        await session.execute(
-            update(ConversationMessage)
-            .where(ConversationMessage.message_id == message_id)
-            .values(vector_status=status)
-        )
-        await session.commit()
 
 
 async def vectorize_message(message: ConversationMessage) -> bool:
@@ -274,17 +254,7 @@ async def scan_pending_messages() -> int:
     offset = 0
 
     while total_pushed < PENDING_SCAN_MAX_TOTAL:
-        # 查询 pending 状态的消息
-        async with AsyncSessionLocal() as session:
-            result = await session.execute(
-                select(ConversationMessage.message_id)
-                .where(ConversationMessage.vector_status == "pending")
-                .where(ConversationMessage.create_time >= cutoff_ts)
-                .order_by(ConversationMessage.create_time.desc())
-                .offset(offset)
-                .limit(PENDING_SCAN_BATCH_SIZE)
-            )
-            message_ids = [row[0] for row in result.fetchall()]
+        message_ids = await _crud_scan_pending(cutoff_ts, offset, PENDING_SCAN_BATCH_SIZE)
 
         if not message_ids:
             break
